job_crawler/requirements.txt
firebase-admin>=6.2.0
python-dotenv>=1.0.0
firecrawl-py>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
apscheduler>=3.10.0 